Refactor CobaltBrowserMainParts into its own file.
This is in preparation for some non-trivial UMA initialization that
has to happen within CobaltBrowserMainParts. It was getting too messy
having this class within cobalt_content_browser_client.cc.

b/372559349
