+++ cobalt/browser/BUILD.gn
@@ -22,6 +22,8 @@ source_set("browser") {
   sources = [
     "cobalt_browser_interface_binders.cc",
     "cobalt_browser_interface_binders.h",
+    "cobalt_browser_main_parts.cc",
+    "cobalt_browser_main_parts.h",
     "cobalt_content_browser_client.cc",
     "cobalt_content_browser_client.h",
     "cobalt_web_contents_observer.cc",
